[docs]:update the eval code
