+++ =0.8.2
@@ -1,0 +1,33 @@
+Looking in indexes: https://pypi.org/simple, https://pypi.ngc.nvidia.com
+Requirement already satisfied: torchvision in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (0.15.2)
+Requirement already satisfied: numpy in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torchvision) (1.22.4)
+Requirement already satisfied: requests in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torchvision) (2.31.0)
+Requirement already satisfied: torch==2.0.1 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torchvision) (2.0.1)
+Requirement already satisfied: pillow!=8.3.*,>=5.3.0 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torchvision) (10.2.0)
+Requirement already satisfied: filelock in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torch==2.0.1->torchvision) (3.13.1)
+Requirement already satisfied: typing-extensions in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torch==2.0.1->torchvision) (4.10.0)
+Requirement already satisfied: sympy in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torch==2.0.1->torchvision) (1.12)
+Requirement already satisfied: networkx in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torch==2.0.1->torchvision) (3.1)
+Requirement already satisfied: jinja2 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torch==2.0.1->torchvision) (3.1.3)
+Requirement already satisfied: nvidia-cuda-nvrtc-cu11==11.7.99 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torch==2.0.1->torchvision) (11.7.99)
+Requirement already satisfied: nvidia-cuda-runtime-cu11==11.7.99 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torch==2.0.1->torchvision) (11.7.99)
+Requirement already satisfied: nvidia-cuda-cupti-cu11==11.7.101 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torch==2.0.1->torchvision) (11.7.101)
+Requirement already satisfied: nvidia-cudnn-cu11==8.5.0.96 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torch==2.0.1->torchvision) (8.5.0.96)
+Requirement already satisfied: nvidia-cublas-cu11==11.10.3.66 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torch==2.0.1->torchvision) (11.10.3.66)
+Requirement already satisfied: nvidia-cufft-cu11==10.9.0.58 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torch==2.0.1->torchvision) (10.9.0.58)
+Requirement already satisfied: nvidia-curand-cu11==10.2.10.91 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torch==2.0.1->torchvision) (10.2.10.91)
+Requirement already satisfied: nvidia-cusolver-cu11==11.4.0.1 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torch==2.0.1->torchvision) (11.4.0.1)
+Requirement already satisfied: nvidia-cusparse-cu11==11.7.4.91 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torch==2.0.1->torchvision) (11.7.4.91)
+Requirement already satisfied: nvidia-nccl-cu11==2.14.3 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torch==2.0.1->torchvision) (2.14.3)
+Requirement already satisfied: nvidia-nvtx-cu11==11.7.91 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torch==2.0.1->torchvision) (11.7.91)
+Requirement already satisfied: triton==2.0.0 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from torch==2.0.1->torchvision) (2.0.0)
+Requirement already satisfied: setuptools in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from nvidia-cublas-cu11==11.10.3.66->torch==2.0.1->torchvision) (68.2.2)
+Requirement already satisfied: wheel in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from nvidia-cublas-cu11==11.10.3.66->torch==2.0.1->torchvision) (0.41.2)
+Requirement already satisfied: cmake in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from triton==2.0.0->torch==2.0.1->torchvision) (3.28.3)
+Requirement already satisfied: lit in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from triton==2.0.0->torch==2.0.1->torchvision) (17.0.6)
+Requirement already satisfied: charset-normalizer<4,>=2 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from requests->torchvision) (3.3.2)
+Requirement already satisfied: idna<4,>=2.5 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from requests->torchvision) (3.6)
+Requirement already satisfied: urllib3<3,>=1.21.1 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from requests->torchvision) (2.2.1)
+Requirement already satisfied: certifi>=2017.4.17 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from requests->torchvision) (2024.2.2)
+Requirement already satisfied: MarkupSafe>=2.0 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from jinja2->torch==2.0.1->torchvision) (2.1.5)
+Requirement already satisfied: mpmath>=0.19 in /data/xiaogeng_liu/miniconda3/envs/opensora/lib/python3.8/site-packages (from sympy->torch==2.0.1->torchvision) (1.3.0)
